Indicate that you have a containerit::write method
Because if the pkg is not attached, the user will not have access to this method, and it will fail with a cryptic error

For example, I like to do this in a workflow, without attaching the pkg:

```
dockerfile <- containerit::dockerfile()
containerit::write(dockerfile, "dockerfile")
```
But if I just do `write(dockerfile, "dockerfile")` it will fail with an error. And some users might get stuck on that. 

Just a thought!